Moves database operations into separate modules.
